# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API server for the Tourist Trip Planning System.",
        "title": "Tourist Trip Planning System API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/itineraries/{id}/activities": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Fetch all activities for a specific itinerary day",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Trips"
                ],
                "summary": "Get itinerary activities",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Itinerary ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ItineraryActivity"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/trips/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Trips"
                ],
                "summary": "Creating a new itinerary",
                "parameters": [
                    {
                        "description": "Travel information",
                        "name": "trip",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TripPlanRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "trip_id",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/trips/{id}/generate-options": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Trips"
                ],
                "summary": "Generation of travel options (with budget types)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Trip ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.TripOption"
                            }
                        }
                    }
                }
            }
        },
        "/api/trips/{id}/itinerary": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Fetch all days of a specific trip's itinerary",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Trips"
                ],
                "summary": "Get trip itinerary",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Trip ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.TripItinerary"
                            }
                        }
                    }
                }
            }
        },
        "/api/trips/{id}/select-option": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Confirm the chosen travel tier and logistics IDs to finalize the plan",
                "tags": [
                    "Trips"
                ],
                "summary": "Finalize trip selection",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Trip ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Selected tier and entity IDs",
                        "name": "selection",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trip finalized successfully",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/users/preferences": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Saving user preferences",
                "parameters": [
                    {
                        "description": "Preferences",
                        "name": "preferences",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserPreferences"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "preference_id",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/users/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "New user registration",
                "parameters": [
                    {
                        "description": "Registration information",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserRegistrationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "user_id",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Flight": {
            "type": "object",
            "properties": {
                "airline": {
                    "type": "string"
                },
                "duration_minutes": {
                    "type": "integer"
                },
                "flight_id": {
                    "type": "integer"
                },
                "from_city_id": {
                    "type": "integer"
                },
                "price": {
                    "type": "number"
                },
                "to_city_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                }
            }
        },
        "models.Hotel": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "city_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "hotel_id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "price_per_night": {
                    "type": "number"
                },
                "rating": {
                    "type": "number"
                },
                "stars": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                }
            }
        },
        "models.ItineraryActivity": {
            "type": "object",
            "properties": {
                "activity_id": {
                    "type": "integer"
                },
                "activity_type": {
                    "type": "string"
                },
                "attraction_id": {
                    "$ref": "#/definitions/sql.NullInt64"
                },
                "created_at": {
                    "type": "string"
                },
                "end_time": {
                    "type": "string"
                },
                "flight_id": {
                    "$ref": "#/definitions/sql.NullInt64"
                },
                "hotel_id": {
                    "$ref": "#/definitions/sql.NullInt64"
                },
                "itinerary_id": {
                    "type": "integer"
                },
                "notes": {
                    "type": "string"
                },
                "order_number": {
                    "type": "integer"
                },
                "restaurant_id": {
                    "$ref": "#/definitions/sql.NullInt64"
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
...
        "models.TripItinerary": {
            "type": "object",
            "properties": {
                "Itinerary_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "day_number": {
                    "type": "integer"
                },
                "notes": {
                    "type": "string"
                },
                "trip_id": {
                    "type": "integer"
                }
            }
        },
        "models.TripOption": {
            "type": "object",
            "properties": {
                "activites_budget": {
                    "type": "number"
                },
                "hotel": {
                    "$ref": "#/definitions/models.Hotel"
                },
                "inbound_flight": {
                    "$ref": "#/definitions/models.Flight"
                },
                "logistics_budget": {
                    "type": "number"
                },
                "more_money": {
                    "type": "number"
                },
                "outbound_flight": {
                    "$ref": "#/definitions/models.Flight"
                },
                "tier": {
                    "type": "string"
                },
                "total_price_of_money": {
                    "type": "number"
                }
            }
        },
        "models.TripPlanRequest": {
            "type": "object",
            "properties": {
                "currency": {
                    "type": "string"
                },
                "destination_city_id": {
                    "type": "integer"
                },
                "end_date": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "total_price": {
                    "type": "number"
                }
            }
        },
        "models.UserPreferences": {
            "type": "object",
            "properties": {
                "budget_max": {
                    "type": "number"
                },
                "budget_min": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "home_city_id": {
                    "type": "integer"
                },
                "preference_id": {
                    "type": "integer"
                },
                "preferred_categories": {
                    "type": "string"
                },
                "travel_style": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.UserRegistrationRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "sql.NullInt64": {
            "type": "object",
            "properties": {
                "int64": {
                    "type": "integer",
                    "format": "int64"
                },
                "valid": {
                    "description": "Valid is true if Int64 is not NULL",
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}